Move numeric bound-related classes under common package
