+++ java/src/main/kotlin/io/spine/validation/java/JavaValidationRenderer.kt
@@ -42,10 +42,10 @@ import io.spine.validation.java.generate.MessageValidationCode
 import io.spine.validation.java.generate.ValidationCodeInjector
 import io.spine.validation.java.generate.option.DistinctGenerator
 import io.spine.validation.java.generate.option.GoesGenerator
-import io.spine.validation.java.generate.option.MaxGenerator
-import io.spine.validation.java.generate.option.MinGenerator
+import io.spine.validation.java.generate.option.bound.MaxGenerator
+import io.spine.validation.java.generate.option.bound.MinGenerator
 import io.spine.validation.java.generate.option.PatternGenerator
-import io.spine.validation.java.generate.option.RangeGenerator
+import io.spine.validation.java.generate.option.bound.RangeGenerator
 import io.spine.validation.java.generate.option.ChoiceGenerator
 import io.spine.validation.java.generate.option.RequiredGenerator
 import io.spine.validation.java.generate.option.ValidateGenerator
